requirements.txt
discord.py[voice]==2.3.2
yt-dlp==2023.7.6
python-dotenv==1.0.0
Flask==2.3.2